Extract Python string tensor utils into a separate crate
