+++ source/carton-bindings-py/Cargo.toml
@@ -17,4 +17,4 @@ semver = {version = "1.0.16"}
 target-lexicon = "0.12.5"
 tokio = { version = "1", features = ["io-util"] }
 async-trait = "0.1"
-widestring = "1.0.2"
+carton-utils-py = {path = "../carton-utils-py"}
